+++ src/LearningObjects/LearningObjectInteractor.spec.ts
@@ -4,7 +4,6 @@ import { MOCK_OBJECTS } from '../tests/mocks';
 import {
   updateObjectLastModifiedDate,
   updateParentsDate,
-  getRecentChangelog,
 } from './LearningObjectInteractor';
 
 const dataStore: DataStore = new MockDataStore(); // DataStore
@@ -28,26 +27,5 @@ describe('Interactor: LearningObjectInteractor', () => {
         date: Date.now().toString(),
       }),
     ).resolves.toBe(undefined);
-  });
-  it('should get latest changelog for a learning object', async () => {
-    try {
-      await expect(getRecentChangelog(
-        dataStore,
-        MOCK_OBJECTS.LEARNING_OBJECT_ID,
-      ))
-      .resolves.toEqual({
-        _id: '1234',
-        learningObjectId: '1223',
-        logs: [
-            {
-                userId: '123',
-                date: '2019-02-06T15:52:10.894Z',
-                text: 'hello',
-            },
-        ],
-    });
-    } catch (error) {
-      console.log(error);
-    }
   });
 });
